+++ src/common/ExtensionInfo.ts
