+++ src/common/Mediator.ts
@@ -2,6 +2,7 @@ import { LayoutCard } from './LayoutCard';
 import { SummaryWord } from './SummaryWord';
 import { WordCard } from './WordCard';
 
+
 export interface Mediator {
   summary: SummaryWord;
   layout: LayoutCard;
