+++ src/common/SummaryWord.ts
