+++ src/common/WordCard.ts
