+++ src/renderer/states/State.ts
@@ -1,6 +1,6 @@
 
-import { ExtensionInfo } from '../ExtensionInfo';
-import { Mediator } from '../Mediator';
+import { ExtensionInfo } from '../../common/ExtensionInfo';
+import { Mediator } from '../../common/Mediator';
 
 import Bookshelf from './Bookshelf';
 import ThemeParameter from './ThemeParameter';
